admin categories

+++ sources/admin_categories.php
@@ -29,7 +29,7 @@
 if ( !defined('INCLUDED') )
 	exit();
 
-$result = $db->query("SELECT id, name, sort_id FROM ".TABLE_PREFIX."cats ORDER BY sort_id ASC, id");
+$result = $db->query("SELECT id, name, sort_id FROM ".TABLE_PREFIX."cats ORDER BY sort_id ASC, name");
 $cats = array();
 $filled_in = true;
 while ( $catinfo = $db->fetch_result($result) ) {
@@ -44,52 +44,56 @@ $_GET['do'] = ( !empty($_GET['do']) ) ? $_GET['do'] : 'index';
 
 if ( $_GET['do'] == 'index' ) {
 	
-	if ( $filled_in ) {
+	$content = '<p>'.$lang['CategoriesInfo'].'</p>';
+	$content .= '<ul id="admincategoriesmenu">';
+		$content .= '<li><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'add')).'">'.$lang['CategoriesAddNewCat'].'</a></li> ';
+		$content .= '<li><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'adjustsortids')).'">'.$lang['CategoriesAdjustSortIDs'].'</a></li> ';
+		$content .= '<li><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'autosort')).'">'.$lang['CategoriesSortAutomatically'].'</a></li> ';
+	$content .= '</ul>';
+	
+	if ( !count($cats) ) {
 		
-		$biggest_sort_id = 0;
-		foreach ( $cats as $cat ) {
-			
-			$db->query("UPDATE ".TABLE_PREFIX."cats SET sort_id = ".$_POST['sort_id-'.$cat['id']]." WHERE id = ".$cat['id']);
-			
-			if ( $_POST['sort_id-'.$cat['id']] > $biggest_sort_id )
-				$biggest_sort_id = $_POST['sort_id-'.$cat['id']];
-			
-		}
-		
-		if ( !empty($_POST['new_cat_name']) ) {
-			
-			$sort_id = ( isset($_POST['new_sort_id']) && valid_int($_POST['new_sort_id']) ) ? $_POST['new_sort_id'] : $biggest_sort_id+1;
-			$db->query("INSERT INTO ".TABLE_PREFIX."cats VALUES (NULL, '".$_POST['new_cat_name']."', ".$sort_id.")");
-			
-			$content = '<p>'.$lang['CategoriesSortChangesNewCatApplied'].'</p>';
-			
-		} else {
-			
-			$content = '<p>'.$lang['CategoriesSortChangesApplied'].'</p>';
-			
-		}
+		$content .= '<p>'.$lang['CategoriesNoCatsExist'].'</p>';
 		
 	} else {
 		
-		$content = ( $_SERVER['REQUEST_METHOD'] == 'POST' ) ? '<p>'.$lang['CategoriesMissingFields'].'</p>' : '<p>'.$lang['CategoriesInfo'].'</p>';
+		if ( $_SERVER['REQUEST_METHOD'] == 'POST' ) {
+			
+			if ( $filled_in ) {
+				
+				foreach ( $cats as $cat ) {
+					
+					$db->query("UPDATE ".TABLE_PREFIX."cats SET sort_id = ".$_POST['sort_id-'.$cat['id']]." WHERE id = ".$cat['id']);
+					$cats[$cat['id']]['sort_id'] = $_POST['sort_id-'.$cat['id']];
+					
+				}
+				
+				$content .= '<p>'.$lang['CategoriesSortChangesApplied'].'</p>';
+				
+			} else {
+				
+				$content .= '<p><strong>'.$lang['CategoriesMissingFields'].'</strong></p>';
+				
+			}
+			
+		}
 		
 		$content .= '<form action="'.$functions->make_url('admin.php', array('act' => 'categories')).'" method="post">';
-		$content .= '<table id="admincatstable"><tr><th>'.$lang['CategoriesCatName'].'</th><th>'.$lang['CategoriesEdit'].'</th><th>'.$lang['CategoriesDelete'].'</th><th>'.$lang['CategoriesSortID'].'</th></tr>';
+		$content .= '<table id="admincatstable"><tr><th>'.$lang['CategoriesCatName'].'</th><th class="action">'.$lang['CategoriesEdit'].'</th><th class="action">'.$lang['CategoriesDelete'].'</th><th class="action">'.$lang['CategoriesSortID'].'</th></tr>';
 		
+		$i = 1;
 		if ( count($cats) ) {
 			
-			$i = 1;
 			foreach ( $cats as $cat ) {
 				
-				$content .= '<tr><td><em>'.unhtml(stripslashes($cat['name'])).'</em></td><td class="action"><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'edit', 'id' => $cat['id'])).'">'.$lang['CategoriesEdit'].'</a></td><td class="action"><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'delete', 'id' => $cat['id'])).'">'.$lang['CategoriesDelete'].'</a></td><td><input type="text" name="sort_id-'.$cat['id'].'" value="'.$cat['sort_id'].'" size="3" maxlength="11" tabindex="'.$i.'" /></td></tr>';
+				$content .= '<tr><td><em>'.unhtml(stripslashes($cat['name'])).'</em></td><td class="action"><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'edit', 'id' => $cat['id'])).'">'.$lang['CategoriesEdit'].'</a></td><td class="action"><a href="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'delete', 'id' => $cat['id'])).'">'.$lang['CategoriesDelete'].'</a></td><td class="action"><input type="text" name="sort_id-'.$cat['id'].'" value="'.$cat['sort_id'].'" size="3" maxlength="11" tabindex="'.$i.'" /></td></tr>';
 				$i++;
 				
 			}
 			
 		}
 		
-		$content .= '<tr><td><input type="text" name="new_cat_name" size="30" maxlength="255" /></td><td class="action"></td><td class="action"></td><td><input type="text" name="new_sort_id" size="3" maxlength="11" /></td></tr>';
-		$content .= '<tr><td colspan="4" class="submit"><input type="submit" value="'.$lang['Save'].'" /> <input type="reset" value="'.$lang['Reset'].'" /></td></tr></table></form>';
+		$content .= '<tr><td colspan="4" class="submit"><input type="submit" value="'.$lang['Save'].'" tabindex="'.$i.'" /> <input type="reset" value="'.$lang['Reset'].'" tabindex="'.($i+1).'" /></td></tr></table></form>';
 		
 	}
 	
@@ -99,7 +103,10 @@ if ( $_GET['do'] == 'index' ) {
 		
 		if ( !empty($_POST['delete']) ) {
 			
-			$admin_functions->delete_forums('cat_id = '.$_GET['id']);
+			if ( !empty($_POST['move_contents']) && valid_int($_POST['move_contents']) && array_key_exists($_POST['move_contents'], $cats) )
+				$db->query("UPDATE ".TABLE_PREFIX."forums SET cat_id = ".$_POST['move_contents']." WHERE cat_id = ".$_GET['id']);
+			else
+				$admin_functions->delete_forums('cat_id = '.$_GET['id']);
 			$db->query("DELETE FROM ".TABLE_PREFIX."cats WHERE id = ".$_GET['id']);
 			
 		}
@@ -108,7 +115,25 @@ if ( $_GET['do'] == 'index' ) {
 		
 	} else {
 		
-		$content = '<h2>'.$lang['CategoriesConfirmCatDelete'].'</h2><p><strong>'.sprintf($lang['CategoriesConfirmCatDeleteContent'], '<em>'.unhtml(stripslashes($cats[$_GET['id']]['name'])).'</em>').'</strong></p><form action="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'delete', 'id' => $_GET['id'])).'" method="post"><p><input type="submit" name="delete" value="'.$lang['Delete'].'" /> <input type="submit" value="'.$lang['Cancel'].'" /></p></form>';
+		$content = '<h2>'.$lang['CategoriesConfirmCatDelete'].'</h2>';
+		$content .= '<p><strong>'.sprintf($lang['CategoriesConfirmCatDeleteContent'], '<em>'.unhtml(stripslashes($cats[$_GET['id']]['name'])).'</em>').'</strong></p>';
+		$content .= '<form action="'.$functions->make_url('admin.php', array('act' => 'categories', 'do' => 'delete', 'id' => $_GET['id'])).'" method="post">';
+		if ( count($cats) >= 2 ) {
+			
+			$category_select = '<select name="move_contents"><option value="">'.$lang['CategoriesDeleteContents'].'</option>';
+			foreach ( $cats as $cat ) {
+				
+				if ( $cat['id'] != $_GET['id'] )
+					$category_select .= '<option value="'.$cat['id'].'">'.unhtml(stripslashes($cat['name'])).'</option>';
+				
+			}
+			$category_select .= '</select>';
+			
+			$content .= '<p>'.sprintf($lang['CategoriesMoveContents'], $category_select).'</p>';
+			
+		}
+		$content .= '<p><input type="submit" name="delete" value="'.$lang['Delete'].'" /> <input type="submit" value="'.$lang['Cancel'].'" /></p>';
+		$content .= '</form>';
 		
 	}
